+++ src/minicraft/level/tile/farming/WheatTile.java
@@ -1,0 +1,8 @@
+package minicraft.level.tile.farming;
+
+public class WheatTile extends Plant {
+
+	public WheatTile(String name) {
+		super(name);
+	}
+}
