+++ src/resources/localization/english_en-us.mcpl
@@ -401,11 +401,14 @@ Sand
 Cactus
 Cactus
 
-Seeds
-Seeds
+Wheat Seeds
+Wheat Seeds
 
 Grass Seeds
 Grass Seeds
+
+Potato
+Potato
 
 Bone
 Bone
